# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reservations": {
      "post": {
        "description": "Applies every business rule and guarantees there is no overlap. Intervals are half-open `[start, end)`: a reservation ending at 11:00 does not clash with one starting at 11:00.",
        "operationId": "ReservationsController_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Repeating the request with the same key returns the reservation already created instead of creating a second one.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          },
          "409": {
            "description": "The slot is already taken (`OVERLAPPING_RESERVATION`, carrying the conflicting range), or a rule forbids it: outside operating hours, resource blocked, or the active reservation limit reached."
          },
          "422": {
            "description": "Invalid range, a range in the past, or attendees above capacity."
          }
        },
        "summary": "Create a reservation",
        "tags": ["Reservations"]
      },
      "get": {
        "description": "Without the admin role only your own are returned, and the userId parameter is ignored. The range returns reservations that **overlap** it.",
        "operationId": "ReservationsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number, starting at 1.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": ["ASC", "DESC"]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field. Each controller restricts the allowed values, because\ninterpolating a client-supplied column name into ORDER BY is an injection\nvector that no parameter binding protects against.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceId",
            "required": false,
            "in": "query",
            "description": "Filter by resource.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user. Ignored without the admin role: you only see your own.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceTypeId",
            "required": false,
            "in": "query",
            "description": "Every reservation for one family of resources.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Repeatable: `?status=CONFIRMED&status=CANCELLED`.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["CONFIRMED", "CANCELLED", "COMPLETED"]
              }
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Returns the reservations that **overlap** the range, not those contained\nwithin it: asking \"what is booked this week\" must include the meeting\nthat started on Friday and ends on Monday.",
            "schema": {
              "example": "2026-09-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-09-30T23:59:59Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ReservationDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "List reservations with filters and pagination",
        "tags": ["Reservations"]
      }
    },
    "/api/v1/reservations/{id}": {
      "get": {
        "operationId": "ReservationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          }
        },
        "summary": "Get one reservation",
        "tags": ["Reservations"]
      },
      "patch": {
        "description": "Re-runs every rule while excluding this reservation, so that it does not collide with the slot it currently occupies.",
        "operationId": "ReservationsController_reschedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          }
        },
        "summary": "Reschedule a reservation",
        "tags": ["Reservations"]
      }
    },
    "/api/v1/reservations/{id}/cancellation": {
      "post": {
        "description": "Idempotent: cancelling an already cancelled reservation returns 200. The slot is freed immediately, because the constraint only indexes CONFIRMED rows.",
        "operationId": "ReservationsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          }
        },
        "summary": "Cancel a reservation",
        "tags": ["Reservations"]
      }
    },
...
  },
  "info": {
    "title": "Shared Resource Booking API",
    "description": "Booking API for shared resources.\n\n**Central rule:** two `CONFIRMED` reservations never overlap on the same resource.\nIntervals are half-open `[start, end)`, so 10:00–11:00 and 11:00–12:00 are\ncontiguous and both valid.\n\nErrors follow RFC 7807 (`application/problem+json`).",
    "version": "0.3.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie issued by POST /api/auth/sign-in/email"
      }
    },
    "schemas": {
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 137
          },
          "totalPages": {
            "type": "number",
            "example": 7
          }
        },
        "required": ["page", "limit", "total", "totalPages"]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Declared explicitly as an untyped array.\n\nSwagger cannot resolve `T` at runtime, and letting the CLI plugin infer\nit makes the schema self-referential -- the app fails to boot with\n\"circular dependency detected (property key: data)\". The real item type\nis composed by ApiPaginatedResponse() below, via allOf.",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": ["data", "meta"]
      },
      "ResourceTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "meeting-room"
          },
          "name": {
            "type": "string",
            "example": "Sala de reuniones"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": ["id", "code", "name", "isActive"]
      },
      "ResourceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "aurora"
          },
          "name": {
            "type": "string",
            "example": "Sala Aurora"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "number",
            "nullable": true,
            "example": 12
          },
          "location": {
            "type": "string",
            "nullable": true,
            "example": "Planta 2"
          },
          "timeZone": {
            "type": "string",
            "description": "IANA zone. The client formats times in this zone, not the browser's.",
            "example": "Europe/Madrid"
          },
          "isActive": {
            "type": "boolean"
          },
          "resourceType": {
            "$ref": "#/components/schemas/ResourceTypeDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "code", "name", "timeZone", "isActive", "createdAt"]
      },
...
      "CreateReservationDto": {
        "type": "object",
        "properties": {
          "startAt": {
            "type": "string",
            "description": "Start in UTC, ISO 8601. **Inclusive.**\n\nMust fall on a multiple of 15 minutes.",
            "example": "2026-09-15T10:00:00Z"
          },
          "endAt": {
            "type": "string",
            "description": "End in UTC, ISO 8601. **Exclusive.**\n\nA reservation ending at 11:00 does not collide with one starting at\n11:00: intervals are half-open `[start, end)`.",
            "example": "2026-09-15T11:00:00Z"
          },
          "attendees": {
            "type": "number",
            "description": "Expected attendees. Checked against the resource's capacity.",
            "minimum": 1,
            "example": 8
          },
          "notes": {
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "description": "The resource to book.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Title shown in the schedule.",
            "maxLength": 160
          }
        },
        "required": ["startAt", "endAt", "resourceId", "title"]
      },
      "ReservationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "startAt": {
            "format": "date-time",
            "type": "string",
            "description": "Start in UTC. Inclusive."
          },
          "endAt": {
            "format": "date-time",
            "type": "string",
            "description": "End in UTC. **Exclusive.**"
          },
          "status": {
            "enum": ["CONFIRMED", "CANCELLED", "COMPLETED"],
            "type": "string"
          },
          "attendees": {
            "type": "number",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          },
          "resource": {
            "$ref": "#/components/schemas/ResourceDto"
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "startAt",
          "endAt",
          "status",
          "userId",
          "resourceId",
          "createdAt"
        ]
      },
      "CancelReservationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Reason for the cancellation. It is recorded.",
            "maxLength": 300,
            "example": "The meeting is postponed"
          }
        }
      },
...
    }
  }
}